Switch to a shared tbdiff library and make this an autotools project.
This commit converts tbdiff to being an autotools-based project. This
means that we now support the usual

  autoreconf -i && ./configure && make && make install

process, plus we provide 'make check' for the tbdiff test suite.

The tbdiff library is now build as a shared library and is also
installed into the system for others to use. The library is
libtool-versioned and ships a pkg-config file (tbdiff-1.pc).

The headers were adjusted so that only tbdiff/tbdiff.h may be included
directly; all others are considered internal. The tbdiff-create and
tbdiff-deploy tools were changed to include this header file.

The tbdiff library is still GPL, not LGPL. We might want to change this
in the future.

Thanks to switching to autotools we now have a way to make releases by
means of 'make dist' and 'make distcheck'. Unfortunately, the latter
currently fails, probably due to something being missing in
tbdiff/Makefile.am.