Add new tbdiff-update command to apply a patch to a system.
This command is written in bash and takes four parameters:

  * a device with a btrfs file system to mount
  * the source subvolume name
  * the target subvolume name
  * the path to a patch file or stream

It then creates the target subvolume as a snapshot of the source
subvolume, applies the patch in there, then creates a <target>-run
snapshot and copies the boot files (vmlinuz, extlinux.conf) to the root
file system. Lastly, it adjusts extlinux.conf to point to the -run
snapshot.