Script to merge and syncronize /etc in different system versions
This commit adds a script to merge and syncronize /etc in different system versions.
The first argument read from command line is the mode, which can be one
of the following:
  - test:   the purpose of this mode is to test some merge cases. It
            receives from the command line v1_dir, vu_dir and v2_dir and
            vt_dir. The meaning of these arguments is explained in the
            script.

  - merge:  merges the user changes in /etc in the run system of the
            version given as argument

  - sync:   syncronizes /etc in all run versions, so that this directory is exactly
            the same as the version given as argument.

This commit also includes an auxiliary script to mount the systems directory in a
give directory given as argument, and some testing folders to use with
the test mode.