Remove test mode and provide hooks to use a test suite.
Tests will now be handled by a test suite in a future commit, so this
mode will not be needed anymore. The test suite will work by replacing
the mounting script by a fake mounting script that points to a directory
with a systems folder.

Also add trap again, now that it is more tested.