Modify the way to use 'patch' in 'baserock-system-config-sync'.
baserock-sytem-config-sync:

Changes here are caused by a bug found in GNU patch
managing the permissions of the files when using the
patch command like:

   patch <file_to_apply_patch> -t -o <output_file>

To reproduce the bug:

  echo foo > file1
  echo bar > file2
  diff -u file1 file2 | patch 1 -t -o file3
  ls -l

You can check that the permissions of 'file3' are different
than the permissions of 'file1' or 'file2'.

To avoid the bug, this patch changes the way we are using
patch, using it as following:

  patch <file_to_apply_patch> -t

Since the output file is not specified, the output file will
be the file in which we want to apply the patch. And
due we cannot specify the output file, we are copying
the file to the destination directory, and then applying
the patch there.

As a consequence of changing the way of using patch, now
'patch' generates an extra file when patching fails. This
file is added in the test suite also in this commit.