Add script to modify the bootloader and manage different parallel OS.
This is part of the upgrades work. With this tool you
can now switch between versions of the OS, remove a
version, list all the versions present in the system, get
the default version and the running version, and deploy
a new system.

All of the above is possible with the following subcommands:
 - list
 - deploy
 - get-default
 - get-running
 - remove
 - set-default

It also activates a bootloader menu to choose a version
to boot. The menu is important to make sure the user can
boot the old OS if the new kernel doesn't work.