baserock-system-config-sync: Force copy /etc/passwd and /etc/group
This change is to ensure the existing users will exist after an
upgrade. Otherwise, if there are merge conflicts when upgrading,
the users will be lost and the root password will be deactivated.

If that happens and the only way to access to the system is
through ssh and the system was rebooted after the upgrade (manually
or automatically) then the system won't be accessible anymore.

This change also means that we can no longer make changes to the
base /etc/passwd or /etc/group in the 'fhs-dirs' chunk without adding
a manual hook to add the new users/groups when upgrading old systems.

In the following link is the email thread where was discussed this issue:

http://vlists.pepperfish.net/pipermail/baserock-dev-baserock.org/2014-March/004581.html