Use sys/xattr from libc if attr/xattr is unavailable
On my system tbdiff failed to build because attr/xattr doesn't
exist, also the attr package was not easy to find.

On searching I discovered that the functionality we use in attr.h
is provided by modern versions of libc, as shown here:
https://bugzilla.kernel.org/show_bug.cgi?id=70141

This patch falls back to using libc sys/xattr.h when attr/xattr.h
is unavailable.