Remove use of typedef to avoid naming conflicts with POSIX
POSIX reserves the use of the underscore T postfix for types,
since most of the time we don't actually need to typedef the types
and since it's more clear to prefix structs and enums with the
struct and enum keywords, I've removed all uses of typedef and
gone with the more simple approach used in the kernel.

One typedef remains which is tbd_cmd_type, since this should be
a native type of a fixed size. The postfix changed to type to show
that it's a type and to ensure it doesn't collide with POSIX.