Actually use atomic_symlink_update for something
This function was added to ensure safe, automatic updates of the default
system symlink, but for some reason, was never used.

Instead it open-coded a vulnerable way of doing this, which could fail
if an upgrade was aborted at the wrong time.

So now we use atomic_symlink_update, which won't fail, and the worst
that can happen is a temporary directory gets left around.