Generate extlinux.conf using deployment options from /baserock/deployment.meta
This will add the initramfs to the config and tell it to mount the disk
by UUID if INITRAMFS_PATH was provided in the initial deployment.

It will also include the extra KERNEL_ARGS provided.

This is required to be able to upgrade a system that uses an initramfs
and have it continue to use the initramfs.

The system will continue to work without this patch if the initramfs'
only responsibility was translating a UUID into a disk path, since
system-version-manager would generate an extlinux.conf that boots with
the same disk as it is currently using, but that would break if the
device enumeration order changed e.g. if another disk was inserted
before the rootfs, or the disk is transplanted into another machine.